+++ src/animations/slideIn.js
@@ -1,8 +1,8 @@
 import { keyframes } from 'styled-components';
 
-const slideIn = distance => keyframes`
-  0% { transform: translateX(${distance}); opacity: 0; }
-  100% { transform: translateX(0); opacity: 1;}
+const slideIn = (distanceX, distanceY = 0) => keyframes`
+  0% { transform: translate(${distanceX}, ${distanceY}); opacity: 0; }
+  100% { transform: translate(0, 0); opacity: 1;}
 `;
 
 export default slideIn;