+++ src/components/DropdownTitle/DropdownTitle.js
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 
 const Wrapper = styled.div`
   padding: 0.5rem 1rem;
-  font-weight: bold;
+  font-weight: 500;
 `;
 
 const DropdownTitle = ({ children, ...otherProps }) => (