+++ src/theme/decorations.js
@@ -1,5 +1,6 @@
 const decorations = {
   shadow: '0 2px 4px 0 rgba(49, 50, 51, 0.2)',
+  borderRadius: '0.25rem',
   noShadow: 'none',
 };